Add better error handling

+++ src/test/resources/xml/corpus/post_to_storage.sh
@@ -7,13 +7,24 @@
 : ${FILE:="illum.xml"}
 
 : ${STORAGE:="http://localhost:8080/ds-storage/v1"}
+: ${ENDPOINT:="$STORAGE/record/createOrUpdateRecord"}
 
 ID="doms.radio:$FILE"
 
 T=$(mktemp)
 echo '{"id":"'$ID'", "base":"doms.radio", "data":'$(jq -R -s '.' < $FILE)'}' > $T
-curl -s -X POST "$STORAGE/createOrUpdateRecord" -H  "accept: */*" -H  "Content-Type: application/json" -d @${T}
+
+STATUSCODE=$(curl -s --output /dev/stderr --write-out "%{http_code}" -X POST "$ENDPOINT" -H  "accept: */*" -H  "Content-Type: application/json" -d @${T})
+
+if [ ! "$STATUSCODE" -eq 204 ]; then
+    >&2 echo "Error: Unable to post content to ds-storage"
+    >&2 echo "Got HTTP code $STATUSCODE for POST to $ENDPOINT"
+    >&2 echo "Check if storage is running at ${STORAGE}"
+    exit 2
+fi
+
+#curl -s -X POST "$STORAGE/record/createOrUpdateRecord" -H  "accept: */*" -H  "Content-Type: application/json" -d @${T}
 rm "$T"
 
 echo "Indexed $ID to ds-storage."
-echo "Access at ${STORAGE}/getRecord?id=$ID"
+echo "Access at ${STORAGE}/record/$ID"
